Move typed EventStream to akka.event.typed package
